🔨 Make Lomax into a windows service on windows
